Add user metadata key checks for oss && refeactor some code
