+++ core/src/raw/http_util/mod.rs
@@ -44,7 +44,6 @@ pub use header::parse_header_to_str;
 pub use header::parse_into_metadata;
 pub use header::parse_last_modified;
 pub use header::parse_location;
-pub use header::parse_metadata;
 
 mod uri;
 pub use uri::percent_decode_path;
