+++ core/src/services/oss/backend.rs
@@ -449,7 +449,8 @@ impl Access for OssBackend {
             StatusCode::OK => {
                 let headers = resp.headers();
                 let mut meta =
-                    parse_metadata(path, constants::USER_METADATA_PREFIX, resp.headers())?;
+                    self.core
+                        .parse_metadata(path, constants::X_OSS_META_PREFIX, resp.headers())?;
 
                 if let Some(v) = parse_header_to_str(headers, "x-oss-version-id")? {
                     meta.set_version(v);
